requirements.txt
gin
jax
jaxlib
trax
matplotlib
discord.py